backend/requirements.txt
Django==5.0.3
psycopg2-binary==2.9.11
dj-rest-auth==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-jazzmin==3.0.0
Pillow==10.2.0
django-summernote==0.8.20.0
django-taggit==5.0.1
openai==2.28.0
requests==2.31.0
pypdf==4.3.1
markdown==3.7
